+++ src/main/commonsComponents/units/template/form/pageNation/index.tsx
@@ -1,5 +1,6 @@
 import styled from "@emotion/styled";
 
+// 페이지네이션 form
 export default function PageNationPage() {
   return <PageNationWrapper>222</PageNationWrapper>;
 }
