+++ src/main/mainComponents/admin/module.list.tsx
@@ -1,4 +1,4 @@
-import AdminBlockPage from "./modules/block";
+import AdminBlockPage from "./modules/block/block.container.tsx";
 
 export const AdminModulesList: { [key: string]: JSX.Element } = {
   Block: <AdminBlockPage />, // 아이피 관리 페이지
